Refaktoroi: Uudelleennimeä ja siirrä luokka omaan tiedostoonsa;
Tarvitaan uudelleenkäytettäväksi kuntarouhinnassa.
